console: fix utf8 display

+++ os/src/sbi.rs
@@ -28,8 +28,8 @@ pub fn set_timer(timer: usize) {
     sbi_call(SBI_SET_TIMER, timer, 0, 0);
 }
 
-pub fn console_putchar(c: usize) {
-    sbi_call(SBI_CONSOLE_PUTCHAR, c, 0, 0);
+pub fn console_putchar(c: u8) {
+    sbi_call(SBI_CONSOLE_PUTCHAR, c as usize, 0, 0);
 }
 
 pub fn console_getchar() -> usize {
